+++ www/common.js
@@ -114,7 +114,8 @@ ln.APP = {
     MAPS_ME: "maps_me",
     CABIFY: "cabify",
     BAIDU: "baidu",
-    TAXIS_99: "taxis_99"
+    TAXIS_99: "taxis_99",
+    GAODE: "gaode"
 };
 
 /**
@@ -136,7 +137,8 @@ ln.APPS_BY_PLATFORM[ln.PLATFORM.ANDROID] = [
     ln.APP.MAPS_ME,
     ln.APP.CABIFY,
     ln.APP.BAIDU,
-    ln.APP.TAXIS_99
+    ln.APP.TAXIS_99,
+    ln.APP.GAODE
 ];
 ln.APPS_BY_PLATFORM[ln.PLATFORM.IOS] = [
     ln.APP.USER_SELECT,
@@ -156,7 +158,8 @@ ln.APPS_BY_PLATFORM[ln.PLATFORM.IOS] = [
     ln.APP.MAPS_ME,
     ln.APP.CABIFY,
     ln.APP.BAIDU,
-    ln.APP.TAXIS_99
+    ln.APP.TAXIS_99,
+    ln.APP.GAODE
 ];
 ln.APPS_BY_PLATFORM[ln.PLATFORM.WINDOWS] = [
     ln.APP.BING_MAPS
@@ -195,6 +198,7 @@ ln.APP_NAMES[ln.APP.MAPS_ME] = "MAPS.ME";
 ln.APP_NAMES[ln.APP.CABIFY] = "Cabify";
 ln.APP_NAMES[ln.APP.BAIDU] = "Baidu Maps";
 ln.APP_NAMES[ln.APP.TAXIS_99] = "99 Taxi";
+ln.APP_NAMES[ln.APP.GAODE] = "Gaode Maps (Amap)";
 
 /**
  * All possible transport modes
@@ -237,6 +241,12 @@ ln.TRANSPORT_MODES[ln.PLATFORM.ANDROID][ln.APP.MAPS_ME] = [
     ln.TRANSPORT_MODE.TRANSIT
 ];
 ln.TRANSPORT_MODES[ln.PLATFORM.ANDROID][ln.APP.BAIDU] = [
+    ln.TRANSPORT_MODE.DRIVING,
+    ln.TRANSPORT_MODE.WALKING,
+    ln.TRANSPORT_MODE.BICYCLING,
+    ln.TRANSPORT_MODE.TRANSIT
+];
+ln.TRANSPORT_MODES[ln.PLATFORM.ANDROID][ln.APP.GAODE] = [
     ln.TRANSPORT_MODE.DRIVING,
     ln.TRANSPORT_MODE.WALKING,
     ln.TRANSPORT_MODE.BICYCLING,
@@ -286,6 +296,12 @@ ln.TRANSPORT_MODES[ln.PLATFORM.IOS][ln.APP.BAIDU] = [
     ln.TRANSPORT_MODE.BICYCLING,
     ln.TRANSPORT_MODE.TRANSIT
 ];
+ln.TRANSPORT_MODES[ln.PLATFORM.IOS][ln.APP.GAODE] = [
+    ln.TRANSPORT_MODE.DRIVING,
+    ln.TRANSPORT_MODE.WALKING,
+    ln.TRANSPORT_MODE.BICYCLING,
+    ln.TRANSPORT_MODE.TRANSIT
+];
 
 /**
  * Apps by platform that support specifying a start location
@@ -304,7 +320,8 @@ ln.SUPPORTS_START[ln.PLATFORM.ANDROID] = [
     ln.APP.MAPS_ME,
     ln.APP.CABIFY,
     ln.APP.BAIDU,
-    ln.APP.TAXIS_99
+    ln.APP.TAXIS_99,
+    ln.APP.GAODE
 ];
 ln.SUPPORTS_START[ln.PLATFORM.IOS] = [
     ln.APP.USER_SELECT,
@@ -320,7 +337,8 @@ ln.SUPPORTS_START[ln.PLATFORM.IOS] = [
     ln.APP.MAPS_ME,
     ln.APP.CABIFY,
     ln.APP.BAIDU,
-    ln.APP.TAXIS_99
+    ln.APP.TAXIS_99,
+    ln.APP.GAODE
 ];
 ln.SUPPORTS_START[ln.PLATFORM.WINDOWS] = [
     ln.APP.BING_MAPS
@@ -339,7 +357,8 @@ ln.SUPPORTS_START_NAME[ln.PLATFORM.ANDROID] = [
     ln.APP.MOOVIT,
     ln.APP.CABIFY,
     ln.APP.BAIDU,
-    ln.APP.TAXIS_99
+    ln.APP.TAXIS_99,
+    ln.APP.GAODE
 ];
 ln.SUPPORTS_START_NAME[ln.PLATFORM.IOS] = [
     ln.APP.USER_SELECT,
@@ -350,7 +369,8 @@ ln.SUPPORTS_START_NAME[ln.PLATFORM.IOS] = [
     ln.APP.MOOVIT,
     ln.APP.CABIFY,
     ln.APP.BAIDU,
-    ln.APP.TAXIS_99
+    ln.APP.TAXIS_99,
+    ln.APP.GAODE
 ];
 
 /**
@@ -367,7 +387,8 @@ ln.SUPPORTS_DEST_NAME[ln.PLATFORM.ANDROID] = [
     ln.APP.MOOVIT,
     ln.APP.CABIFY,
     ln.APP.BAIDU,
-    ln.APP.TAXIS_99
+    ln.APP.TAXIS_99,
+    ln.APP.GAODE
 ];
 ln.SUPPORTS_DEST_NAME[ln.PLATFORM.IOS] = [
     ln.APP.USER_SELECT,
@@ -380,7 +401,8 @@ ln.SUPPORTS_DEST_NAME[ln.PLATFORM.IOS] = [
     ln.APP.MOOVIT,
     ln.APP.CABIFY,
     ln.APP.BAIDU,
-    ln.APP.TAXIS_99
+    ln.APP.TAXIS_99,
+    ln.APP.GAODE
 ];
 
 /**